server/requirements.txt
fastapi==0.82.0
lightgbm==3.3.2
pandas==1.3.4
python-dotenv==0.21.0
uvicorn==0.12.3
aiofiles==0.6.0
python-multipart==0.0.5
jinja2==2.11.2
gunicorn
uvloop
httptools